TASK: Move `CommandHandlingDependencies` to CommandHandler namespace
