+++ Neos.ContentRepository.Core/Classes/Feature/SubtreeTagging/SubtreeTagging.php
@@ -14,7 +14,7 @@ namespace Neos\ContentRepository\Core\Feature\SubtreeTagging;
  * source code.
  */
 
-use Neos\ContentRepository\Core\CommandHandlingDependencies;
+use Neos\ContentRepository\Core\CommandHandler\CommandHandlingDependencies;
 use Neos\ContentRepository\Core\DimensionSpace;
 use Neos\ContentRepository\Core\EventStore\Events;
 use Neos\ContentRepository\Core\EventStore\EventsToPublish;
