+++ Neos.ContentRepository.Core/Classes/Projection/ContentGraph/NodeAggregate.php
@@ -38,16 +38,13 @@ use Neos\ContentRepository\Core\SharedModel\Workspace\WorkspaceName;
  * The system guarantees the following invariants:
  *
  * - Inside a NodeAggregate, each DimensionSpacePoint has at most one Node which covers it.
- *   To check this, the ReadableNodeAggregateInterface is used (mainly in constraint checks).
+ *   To check this, this class is used (mainly in constraint checks).
  * - The NodeType is always the same for all Nodes in a NodeAggregate
  * - all Nodes inside the NodeAggregate always have the same NodeName.
  * - all nodes inside a NodeAggregate are all of the same *classification*, which can be:
  *   - *root*: for root nodes
  *   - *tethered*: for nodes "attached" to the parent node (i.e. the old "AutoCreatedChildNodes")
  *   - *regular*: for all other nodes.
- *
- * This interface is called *Readable* because it exposes read operations on the set of nodes inside
- * a single NodeAggregate; often used for constraint checks (in command handlers).
  *
  * @api Note: The constructor is not part of the public API
  */
